+++ alacritty/src/input/mod.rs
@@ -646,6 +646,9 @@ impl<T: EventListener, A: ActionContext<T>> Processor<T, A> {
                 _ => ClickState::Click,
             };
 
+            // Don't launch hints if there is a selection active.
+            self.ctx.mouse_mut().block_hint_launcher = !self.ctx.selection_is_empty();
+
             // Load mouse point, treating message bar and padding as the closest cell.
             let display_offset = self.ctx.terminal().grid().display_offset();
             let point = self.ctx.mouse().point(&self.ctx.size_info(), display_offset);
@@ -663,9 +666,6 @@ impl<T: EventListener, A: ActionContext<T>> Processor<T, A> {
 
         match self.ctx.mouse().click_state {
             ClickState::Click => {
-                // Don't launch URLs if this click cleared the selection.
-                self.ctx.mouse_mut().block_hint_launcher = !self.ctx.selection_is_empty();
-
                 self.ctx.clear_selection();
 
                 // Start new empty selection.
@@ -708,7 +708,9 @@ impl<T: EventListener, A: ActionContext<T>> Processor<T, A> {
 
         // Trigger hints highlighted by the mouse.
         let hint = self.ctx.display().highlighted_hint.take();
-        if let Some(hint) = hint.as_ref().filter(|_| button == MouseButton::Left) {
+        if let Some(hint) =
+            hint.as_ref().filter(|hint| Some(button) == hint.mouse_button())
+        {
             self.ctx.trigger_hint(hint);
         }
         self.ctx.display().highlighted_hint = hint;